requirements.txt
Django==4.2.7
celery==5.3.6
redis==5.0.1
psycopg==3.1.12
psycopg2-binary==2.9.1
djangorestframework==3.14.0
drf-api-logger==1.1.12
drf-yasg==1.21.7
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
